+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- **APE**: The default track/disk number is now `0` to line up with ID3v2.
+           This is only used when `set_{track, disk}_total` is used without a corresponding `set_{track, disk}`.
+
+## Fixed
+- **APE**: Track/Disk number pairs are properly converted when writing ([issue](https://github.com/Serial-ATA/lofty-rs/issues/159)) ([PR](https://github.com/Serial-ATA/lofty-rs/pull/216))
+
 ## [0.14.0] - 2023-06-08
 
 ### Added